Add files via upload
This is a code for training a 10-class image classification model using the InceptionV3 architecture. The code uses Tensorflow with the Keras API. The data set is processed using the ImageDataGenerator and split into training and validation subsets. The training process includes the compilation of the model with an SGD optimizer and categorical crossentropy loss function. The early stopping and model checkpoint callbacks are defined to save the best model during training. The last few layers of the base model (InceptionV3) are unfrozen and the model is trained for 30 epochs. The summary of the model and the training history are printed after training.